+++ src/util/DistanzaComparator.java
@@ -6,8 +6,8 @@ import java.util.Comparator;
 public class DistanzaComparator implements Comparator<StellaGr> {
     @Override
     public int compare(StellaGr s1, StellaGr s2) {
-        Float d1 = s1.getDistanza();
-        Float d2 = s2.getDistanza();
+        Double d1 = s1.getDistanza();
+        Double d2 = s2.getDistanza();
 //        if (d1 > d2)
 //            return 1;
 //        if (d1 < d2)
